+++ Momen-API/API/Controllers/VendorsController.cs
@@ -30,8 +30,14 @@ namespace API.Controllers
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<ActionResult<VendorForGetDTO>> Post(VendorForAddDTO model)
         {
-            if (await _vendorRepository.IsExist(model.Name,model.CategoryId).ConfigureAwait(true))
+            if (await _vendorRepository.IsExist(model.Name, model.CategoryId).ConfigureAwait(true))
                 return Conflict(new ApiResponse(409, StringConsts.EXISTED));
+
+            if (await _vendorRepository.IsExistByPhone(model.Phone).ConfigureAwait(true))
+                return Conflict(new ApiResponse(409, StringConcatenates.Exist("Phone", model.Phone)));
+
+            if (await _vendorRepository.IsExistByEmail(model.Email).ConfigureAwait(true))
+                return Conflict(new ApiResponse(409, StringConcatenates.Exist("Email", model.Email)));
 
             Vendor vendor = _mapper.Map<Vendor>(model);
             await _vendorRepository.AddAsync(vendor).ConfigureAwait(true);
@@ -42,8 +48,14 @@ namespace API.Controllers
 
         [HttpPatch("{id:int}/licensefile")]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult> UploadLicenseFile([FromForm] VendorForLicenseDTO model)
+        public async Task<ActionResult> UploadLicenseFile(int id, [FromForm] VendorForLicenseDTO model)
         {
+            if (id != model.Id)
+                return BadRequest(new ApiResponse(400, StringConcatenates.NotEqualIds(id, model.Id)));
+
+            if (!await _vendorRepository.IsExist(id).ConfigureAwait(true))
+                return NotFound(new ApiResponse(404, StringConcatenates.NotExist(id)));
+
             FileOperations.WriteFile("Vendor/License", model.Id, model.File);
             Vendor vendor = await _vendorRepository.GetAsync(model.Id).ConfigureAwait(true);
             vendor.LicenseFileName = model.File.FileName;
@@ -55,8 +67,14 @@ namespace API.Controllers
 
         [HttpPatch("{id:int}/logofile")]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult> UploadLogoFile([FromForm] VendorForLogoDTO model)
+        public async Task<ActionResult> UploadLogoFile(int id, [FromForm] VendorForLogoDTO model)
         {
+            if (id != model.Id)
+                return BadRequest(new ApiResponse(400, StringConcatenates.NotEqualIds(id, model.Id)));
+
+            if (!await _vendorRepository.IsExist(id).ConfigureAwait(true))
+                return NotFound(new ApiResponse(404, StringConcatenates.NotExist(id)));
+
             FileOperations.WriteFile("Vendor/Logo", model.Id, model.File);
             Vendor vendor = await _vendorRepository.GetAsync(model.Id).ConfigureAwait(true);
             vendor.LogoFileName = model.File.FileName;
@@ -68,8 +86,14 @@ namespace API.Controllers
 
         [HttpPatch("{id:int}/personalidfile")]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult> UploadPersonalIdFile([FromForm] VendorForPersonalIdDTO model)
+        public async Task<ActionResult> UploadPersonalIdFile(int id, [FromForm] VendorForPersonalIdDTO model)
         {
+            if (id != model.Id)
+                return BadRequest(new ApiResponse(400, StringConcatenates.NotEqualIds(id, model.Id)));
+
+            if (!await _vendorRepository.IsExist(id).ConfigureAwait(true))
+                return NotFound(new ApiResponse(404, StringConcatenates.NotExist(id)));
+
             FileOperations.WriteFile("Vendor/PersonalId", model.Id, model.File);
             Vendor vendor = await _vendorRepository.GetAsync(model.Id).ConfigureAwait(true);
             vendor.PersonalIdFileName = model.File.FileName;
@@ -89,8 +113,14 @@ namespace API.Controllers
             if (!await _vendorRepository.IsExist(id).ConfigureAwait(true))
                 return NotFound(new ApiResponse(404, StringConcatenates.NotExist(id)));
 
-            if (await _vendorRepository.IsExist(model.Id, model.Name,model.CategoryId).ConfigureAwait(true))
+            if (await _vendorRepository.IsExist(model.Id, model.Name, model.CategoryId).ConfigureAwait(true))
                 return Conflict(new ApiResponse(409, StringConsts.EXISTED));
+
+            if (await _vendorRepository.IsExistByPhone(id, model.Phone).ConfigureAwait(true))
+                return Conflict(new ApiResponse(409, StringConcatenates.Exist("Phone", model.Phone)));
+
+            if (await _vendorRepository.IsExistByEmail(id, model.Email).ConfigureAwait(true))
+                return Conflict(new ApiResponse(409, StringConcatenates.Exist("Email", model.Email)));
 
             Vendor oldVendor = await _vendorRepository.GetAsync(id).ConfigureAwait(true);
             Vendor vendor = _mapper.Map<Vendor>(model);
